fix: refactor types code and remove tx confirmation
